Update node details during node registration
If register-nodes is run a second time with the same set of nodes,
it will blindly try and re-register them, leading to failures from
the Nova or Ironic API. Instead, we now look for identifying marks
of registered nodes, and then update those nodes.

At the end, we check to see if there are any extra nodes registered,
and warn, possibly cleaning them up.

Change-Id: I8652c25077246626810f186259adbf14764f4571