Do not create the _member_ role
Keystone will automatically create the _member_ role if it does not
exist, so there's no need to do it in os-cloud-config. Further,
explicitly creating it will cause an issue when you go to create the
first user in Keystone.

Keystone will try to add the user to the _member_ role, but use the
default *id*:
http://git.openstack.org/cgit/openstack/keystone/tree/keystone/assignment/core.py#n261

Since os-cloud-config has precreated the role, it has a different id, so
then Keystone will try and create _member_ again, and failing with
a duplicate entry.

This change was introduced in os-cloud-config via commit:
Ia4d965335f900e03465f45c701f0bd5a3a4a0ea6

Instead of just reverting that commit, I figured we could remove the
_member_ role creation entirely.

We aren't seeing this failure scenario in CI, because we haven't had a
os-cloud-config release in a while apparently and we don't use
os-cloud-config from git (for better or worse).

Change-Id: Iaa05c71173140e68e47357ac8a9ab3fdb276c838