Tweak commas for clarity
As written (lacking the Oxford Comma), the readme suggests to me that we
had two common bring-up tasks:

* initializing keystone
* configuring Neutron from different code bases

I believe that the actual case is that this is intending to describe two
common bring-up tasks:

* initializing keystone
* configuring Neutron

each of which may be called from different code-bases.

I think that my proposed change makes this clearer.

Change-Id: I7c6e4b219468f92e31fc51b3d1aac8bafd9640d8