Deprecate PKI setup in init-keystone
The initialize() method in os_cloud_config.keystone has a flag to
configure if PKI setup is performed or not. Since devtest is moving
to another script to perform PKI setup, we can deprecate doing it
in init-keystone, and also add options to control if it run or not
from the command line.

Change-Id: Iefc9b367c9844342a42ddf055b35bb7c242742b9