If one node is locked, keep trying the rest.
Prior to this change, if register-nodes encountered a single locked
node, it would abort the whole run - meaning that you can't update the
cloud if even a single node is unavailable.

This change makes register-nodes keep going and update all the nodes
that it can. This makes it idempotent and eventually-consistent - as
nodes free up they'll get caught by later runs.

Change-Id: If9ac06aaea94a93e18be0addbf56ad224de0a5b0
closes-bug: #1382311