CORE-877 Adds capabilities to Ironic nodes
If present in baremetal.json, the given capabilities will be added to
the nodes. This gives the ability to target particular nodes by creating
flavors which demand those capabilities.

Depends informally on I0855f6095704fe9da3cb337908957b500de542d7 which
adds capabilities to the baremetal.json file in the first place (but is
in a different repo).

Upstream-ref: None
Change-Id: I9d4eb625f3fd383dde18dd1ba4117c88e59c0305