Grant admin role to admin user with default domain
For keystone v3 operations, a new grant is needed for admin user. With
this change, a grant of admin role with default domain is created for
admin user, in addition to existing grant of admin role with admin
tenant. Since domain related operations are only supported by v3
keystoneclient, a new v3 client variable introduced and role grant
operations are converted to v3.

Change-Id: I9489e1eddcabf38faf9955cf0424b1283d6707ae