Add support for creating Keystone v3 clients
Due to the Keystone v2 client not supporting domains, we should add
support for using the Keystone v3 client for some operations. This
branch does not change any of the existing Keystone callsites, merely
adds support for creating the client.

Drive-by correcting the method name for the Neutron client tests.

Change-Id: Ifd94b73684df99e3a2ef9ced9b8b4a748ac18416