Do not use https for keystone v2 admin
When passing an SSL hostname for keystone v2 admin access, the code
assumes that v2 admin services will be https. However, that is not the
case and usually is avoided. Since v2 is frozen and will be deprecated
some day, adding SSL for this piece of it seems like extra work for
not much gain. So we just use http.

Change-Id: I9fe4f57dc39e709c8f7549913ac3b1ef1eec81b5