Add support to skip either kernel or ramdisk
Currently, glance.create_or_find_kernel_and_ramdisk() will raise an
exception if either the kernel or ramdisk are not found in Glance.
This makes it fairly awkward for others to use the API if they need
to deal with a missing kernel, but an existing ramdisk since the
function will never check, because the exception is raised when the
kernel isn't found.

Change-Id: I872a080846c7a191039e89afaa587cafbd853fe4