Add support for image names to setup-flavors
To bring setup-flavors into line with register-nodes, and to enact
the plan of putting setup-baremetal on a radical diet, add support
to look up the IDs of the images registered in Glance by passing
names. For backwards compatibility, add only long options, and leave
the short options (and their existing long options) with the existing
behaviour.

Change-Id: I7bead49bbcf0ee06d7bedfbf73987b5d36fbf1b1