Shift EnvironmentVariable fixture setup to base
While investigating a different branch, I came across a test failure
due to the ROOT_DISK environment variable leaking into a test. The
CMDEnvironmentTest already had code to deal with that case, added to
deal with the same problem affecting it, so I moved the code into
the base module so all tests can benefit.

Change-Id: I7e82584f75a309c4cea6fb3f7c0c420358c778e1