Don't register https endpoints for non-ssl services
The current code always registers endpoints as https if there is a
public_endpoint set.  For services that don't currently support ssl,
that results in an https url pointing at a non-https port.  This
change checks for both a public_host and ssl_port setting and only
uses ssl if both are present.

Change-Id: Ibeb038eaaf7faea90c2881834c5d675759cedbb3