register-nodes: add support for node capabilites
This patch makes it possible to configure "capabilities"
when registering nodes in the Ironic API. The ability
to register arbitrary capabilities makes it possible to
more easily deploy on heterogeneous sets of bare metal
where you want to control which roles run on which
hardware (by using the compute capabilities filter in Nova).

Previously it was possible to add these capabilities
manually using 'ironic node-update <uuid> add properties...'.
Using the CLI works fine for ad-hoc settings but it makes
sense to allow users to easily assign and manage capabilites
when nodes are created (or updated) with register-nodes
JSON files as well.

Change-Id: Iffd0274dadaf82a8913baa1c9e06e8478716ed6d