register-nodes: add iboot support for node updates
Updates register-nodes so that it supports node updates when
using the iboot driver. When updating instances for most drivers
we base search for the existing Ironic UUID by the pm_addr
(which is usually an IPMI address or something). For iboot
PDU's the same IP can be shared for multiple compute nodes
so we use a combination of the pm address, relay ID, and port
to uniquely identify each node for the update.

Also, temporarily ignoring C901 (complexity) with this
commit until we can refactor things to remove some
of the deprecated nova baremetal code.

Change-Id: I8e058e0822f5306c2704d737ebdde23271fc08be