Don't try to connect to keystone admin via ssl
We don't currently support running the keystone admin endpoint
behind ssl (see _create_keystone_endpoint, which can only register
public endpoints as ssl), so it doesn't make sense to try to connect
to the admin endpoint via ssl.

For v3, this should be fixed at some point in the future, but in the
meantime let's get the code into a consistent state so it's usable
with public ssl endpoints.  According to the inline comments, the
v2 endpoint is never available via ssl anyway so this is actually
the correct behavior there anyway.

Change-Id: Ie0973bf6d25c837cdef7b0f9aaf4d4b96c3fbff7