Fix nodes registering with fake_pxe driver
The fake_pxe driver doesn't have any power control system to manage the
power state of the nodes so that things like pm_addr, pm_password, and
pm_user are not required for the driver. This patch fixes the assumption
in the os-cloud-config that those information will be present on the
json file when registering nodes.

Closes-Bug: #1495448
Change-Id: Ife39e1d81fbe0761fb843c59fd4787357489cf48