Change nova service's name for v3 to novav3.
A common convention for naming nova service's name for v3 endpoint is to
use novav3/computev3.  This patch enforce this convention.

It also unlocks a blocker for this patch
https://review.openstack.org/#/c/234704 where puppet is unable to handle
services only different by type.  This is being fixed here
https://review.openstack.org/#/c/238954/ but as it's a backward
incompatible change it may take some time to get to master.

This also fixes the red-hat bugzilla.

Change-Id: Ic5997b237ebe052b6db08508cd7c1e85b2093a1d
Closes-Bug: rhbz#1236989