Add wol ironic driver support
The basic pxe_wol driver in ironic allows ironic to be used with
even the most basic hardware setups. This patch allows os-cloud-config
to support importing nodes that use the pxe_wol driver. Typically this
would only be utilised for home lab and poc setups, but significantly
expands the hardware that projects like TripleO can utilise.

Change-Id: Ia47cce83c111b08f10f01f8587de251cc4177baf