Allow control over the Keystone adminurl endpoint
We announce the keystone /v2.0 api in the endpoint by default and the
keystone v2.0 client will use the adminurl by default not the publicurl
so we need to allow more granular control over the adminurl.

We should also not pass the 'public' argument when doing the pki
initialization or when creating the keystone clients because 'public'
is meant to be used as value for the public endpoint (as also the
param description suggests) and not as the host against which connecting
to do the pki (via ssh) or endpoints (via http) setup, that is the
'host' param instead.

Change-Id: I6d6ad06441514c6eca9922937dba9b521dad06db