Allow control over the Keystone adminurl and internalurl endpoint
We publish the keystone /v2.0 api in the endpoints by default and the
keystone v2.0 client will use the adminurl by default, not the publicurl,
so we need to allow a more granular configuration of the adminurl.

We also want the internalurl to stay on an internal network which is not
necessarily the same of either the public or the admin network so we need
to allow customization of the internalurl as well.

Finally we should not pass the 'public' argument when doing the pki
initialization or when creating the keystone clients because 'public'
is meant to be the value to use for the public endpoint (as also the
param description suggests) and not the host against which connecting
to do the pki (via ssh) or endpoints (via http) setup, that is the
'host' param instead.

Change-Id: I6d6ad06441514c6eca9922937dba9b521dad06db