add aodh and gnocchi to keystone service list
This patch adds default values for aodh and gnocchi service/endpoint
to keystone service list to make those services available in TripleO
setup phase.

Note: the list is to have default values and not all services will be
created automatically.

Note: alarming function in ceilometer is deprecated by aodh in Liberty
and will be removed in Mitaka. So, we need aodh to have equivalent
functionality of existing ceilometer.

Change-Id: Ib6c9c4c35da3fb55e0ca8e2d5a58ebaf4204d792