Set compute API version to 2.1, don't use v3
Microversions since Nova API v2.1 are aimed to replace the v3 work. The
/v2.1 is backwards compatible with the legacy /v2 endpoint. What we
called in the past /v3 is now something defunct in-tree. The /v2.1 API
is based on the v3 work, but there are many things that differ, in
particular with the backwards-compat thing. We keep the /v2 path in
api-paste.ini for making sure an upgrade doesn't trample operators and
users but if you look in tree, that's redirecting to the v2.1
codepath (just not asking for microversions). In summary, we only need
one endpoint, ie. /v2.1.

Additional information at https://bugzilla.redhat.com/show_bug.cgi?id=1291291

Closes-Bug: #1564372
Depends-On: Icdda1e08dd691aeca693725931709dcdc0efaf57
Change-Id: I6d64b8bcd0f79f1f298ddc809e6d92fbc2985c45