Add support for registering nodes with the libvirt driver
The pxe_libvirt_{iscsi, agent} driver is part of the
ironic-staging-drivers project and is used to power control and manage
VMs to simulate baremetal machines (similar to what the pxe_ssh driver
does).

This patch is adding support for registering nodes with such driver.
FWIW, RDO already have ironic-staging-drivers packaged.

Change-Id: I5e505e585216ecf123d125db3a711ccba531ad85